Hibernate after three hours of suspend
Maintain the normal `suspend` behavior if the laptop is plugged in.
The dedicated swap partition that allows this to work is only 1/2 the
size of RAM, but it should still be successful most of the time due to
compression and the kernel's best effort to keep the `image_size` to 2/5
of the available RAM by default.

The power button normally would just `poweroff` on press (which I've
done accidentally more than once), and a long press will always be
handled as a hard shutdown by the hardware.

See:
- https://www.kernel.org/doc/html/latest/admin-guide/pm/sleep-states.html?highlight=image_size#basic-sysfs-interfaces-for-system-suspend-and-hibernation