Add bindings for reboot and poweroff
I like having the simple press of the Favorites button to lock the
screen since I use that most often, however it would be nice to not
have to type in the commands in a terminal to reboot or poweroff. The
hardware power button already is set to suspend-then-hibernate, but it
will no longer poweroff gracefully.