Trigger suspend-then-hibernate when idling
This makes sure that it will happen even if the machine has just been
idling (and not just from a lid close or power button press). I reverted
the change to prevent hibernation when we're plugged in, as it's just
easier mentally to have the behavior be consistent no matter the
trigger.

I also switched the ordering or the other idle events so the display
will turn off before locking. It should save a bit more power and also
serve as an indicator that a lock is coming soon.