Allow the EFI boot loader to set its own resolution
So far I've not been able to get the text to be bigger unless using
the "max" value, but then the entries are placed in a weird way on the
screen. The list of generations is so small on the HiDPI screen, and
I'm hoping that once I can clear up the older generations, it will
auto-adjust, but I may just have to live with it.

See:
- https://nixos.org/manual/nixos/unstable/options.html#opt-boot.loader.systemd-boot.consoleMode