Use absolute paths for common lint/fmt commands
I've been experimenting with some transparent Docker-ized wrappers for commands,
and they end up being slow and difficult when running within an editor since it
creates a /tmp file to make the process atomic. Those files don't get mounted
within thing containers, so I'd like to bypass the wrappers and use the system
binaries.

I haven't done this across the board, as most of the time I'll probably bring
these other packages in via a development shell, so they're not always available
in the environment.

I'm also not 100% sure that this is the best way to inject the outPath, but I'm
having trouble finding the right terms to search for to find other examples.