Use custom icons for lf file listings
This unfortunately cannot be configured via a configuration file, so we have to
explicitly set an environment variable with the desired icons. I'm not 100% sure
what all of the other abbreviations correlate with in the interface, but this
seems to cover most things I've run across through some simple navigation.