Create a simple left/right prompt
This was inspired by Kovid Goyal's prompt setup that I've seen in his kitty
demonstration videos. I ended up copying/pasting a lot of the code from the
"scales" built-in example prompt.

I used to have a Starship prompt setup, but I never relied on the complexity
of all of the environmental differences. The only things I really want are:

- the current working directory
- command error code when there's a non-standard exit
- duration for long-running commands
- basic Git branch status info

The `fish_postexec` function is there to add a newline after command output
before the next prompt, but it knows enough to not run when a new terminal is
just opened.

See:
- https://sw.kovidgoyal.net/kitty/changelog/#a-demo-video
- https://github.com/fish-shell/fish-shell/blob/master/share/tools/web_config/sample_prompts/scales.fish