Switch to using Zoom via Flatpak
I'm going to immediately revert this, as it did not work any better due to
external reasons, but I thought I'd document the process.

Sharing works in the browser since it's using `xdg-desktop-portal-wlr`
instead of the private D-Bus APIs that Gnome has since removed access to,
but it's not super efficient on resources.

It sounds like Zoom has officially acknowledged the Wayland issues upstream,
and a fix should be released "soon" (maybe with a 6.0 release), even though
it's been years:

  https://community.zoom.com/t5/Meetings/Wayland-screen-sharing-broken-with-GNOME-41-on-Fedora-35/m-p/40650/highlight/true#M20106

I ended up filing a support ticket to add to the pile of complaints, but it's
busted pretty much everywhere. Their hard-code check of `/etc/os-release`
is also problematic. I'll probably limp along with the web client when I
need to share, and hope it actually gets resolved soon. Alternatively, I
may try using OBS Studio to present my desktop as a virtual camera, which
I believe _can_ be shared. What a kludge.