Split up home configs into common and host-specific
This is mostly a distinction between CLI apps that should be installed
universally, and GUI apps that would only exist on a local machine. I
can abstract this a bit more once additional snowflake machines are under
management.

Note that I wasn't able to figure out how to do conditional imports without
causing infinite recursion, so the home-manager module still imports the
host-specific configuration directly (which then imports common), instead of
defaulting to the common config and then conditionally adding the host-specific
config if it exists. At least I was able to utilize the `networking.hostName`
variable and not have that hard-coded.